Fixed hiding tab renamer when clicking other stuff and when opening a new tab. Fixed compiler warning on Mac.

+++ src/Contrivance/Contrivance_Console.cpp
@@ -62,7 +62,7 @@ void Console::print(const QString& message, ConsoleMessageType type)
 
 void Console::addOutput(const QString& message, const QColor& color, QListWidget* list)
 {
-	if (list->count() == _maximum_log_entries) {
+    if ((unsigned int)list->count() == _maximum_log_entries) {
 		delete list->item(0);
 	}
